+++ app/layout.tsx
@@ -13,7 +13,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang='en' theme='dark'>
+    <html lang='en' data-theme='dark'>
       <body>{children}</body>
     </html>
   );
