+++ app/page.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import Image from 'next/image';
-import Link from 'next/link';
 import Navbar from '@/components/navbar';
 import About from '@/components/about';
 import Contact from '@/components/contact';
@@ -9,13 +7,9 @@ import ProjectGallery from '@/components/projectGallery';
 import AnimatedTitle from '@/components/animatedTitle';
 import { projects } from '@/data/projectData';
 import React, { useRef } from 'react';
+import LaptopAnimation from '@/components/laptopAnimation';
 
 export default function Home() {
-  const ref = useRef(null);
-  React.useEffect(() => {
-    import('@lottiefiles/lottie-player');
-  });
-
   return (
     <div id={'about'}>
       <Navbar />
@@ -23,16 +17,7 @@ export default function Home() {
         <div className='h-10'></div>
         <AnimatedTitle text={'MARK HANLEY'} replay={true} />
         <h2>WEB DEVELOPER</h2>
-        <div className='mx-auto w-9/12 max-w-md'>
-          <lottie-player
-            id='laptop-animation'
-            ref={ref}
-            autoplay
-            loop
-            mode='normal'
-            src='/animations/laptopAnimation.json'
-          ></lottie-player>
-        </div>
+        <LaptopAnimation />
         <main className='stack-lg text-justify'>
           <About heading='About' />
           <ProjectGallery heading='Projects' projects={projects} />
